+++ indexing-rdf/src/main/java/org/qi4j/index/rdf/callback/SingleQualifiedIdentityResultCallback.java
@@ -24,7 +24,7 @@ public class SingleQualifiedIdentityResultCallback
 {
     private EntityReference entityReference;
 
-    public boolean processRow(int row, EntityReference entityReference)
+    public boolean processRow(long row, EntityReference entityReference)
     {
         this.entityReference = entityReference;
         return false;
